Rename containers in users-permissions
