+++ packages/plugins/users-permissions/admin/src/pages/EmailTemplates/utils/schema.js
