+++ packages/plugins/users-permissions/admin/src/pages/Providers/utils/createProvidersArray.js
